move the best parts of sct_utils into spinalcordtoolbox.utils
