+++ spinalcordtoolbox/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python
+
+from .utils import __version__, __sct_dir__, __data_dir__
+
